requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asgiref=3.5.2=py310haa95532_0
bzip2=1.0.8=he774522_0
ca-certificates=2022.10.11=haa95532_0
certifi=2022.9.24=py310haa95532_0
django=4.1=py310haa95532_0
djangorestframework=3.14.0=pypi_0
dnspython=2.2.1=pypi_0
faker=15.1.1=pypi_0
libffi=3.4.2=hd77b12b_4
openssl=1.1.1q=h2bbff1b_0
pip=22.2.2=py310haa95532_0
pymongo=4.3.2=pypi_0
python=3.10.6=hbb2ffb3_1
python-dateutil=2.8.2=pypi_0
python-tzdata=2021.1=pyhd3eb1b0_0
pytz=2022.5=pypi_0
setuptools=65.4.0=py310haa95532_0
six=1.16.0=pypi_0
sqlite=3.39.3=h2bbff1b_0
sqlparse=0.2.4=pypi_0
tk=8.6.12=h2bbff1b_0
tzdata=2022e=h04d1e81_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py310haa95532_2
xz=5.2.6=h8cc25b3_0
zlib=1.2.13=h8cc25b3_0
